requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
python-oracledb>=2.0.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.12.0
sse-starlette>=2.0.0
